Rename SimpleBound -> Bound + add unbind function
